* fix: when the installation is aborted the @DO_NOT_CHANGE@ version string was replaced permanently * tweak: splitted hid-xpadneo into header and code
