+++ hid-xpadneo/dkms.conf
@@ -1,5 +1,5 @@
 PACKAGE_NAME="hid-xpadneo"
-PACKAGE_VERSION="0.6.0"
+PACKAGE_VERSION="@DO_NOT_CHANGE@"
 BUILT_MODULE_NAME[0]="hid-xpadneo"
 BUILT_MODULE_LOCATION[0]="src"
 DEST_MODULE_LOCATION[0]="/kernel/drivers/hid"
