+++ install.sh
@@ -9,22 +9,16 @@ if [[ $EUID != 0 ]]; then
 fi
 
 VERSION=$(cat VERSION)
-
-# backup original files, preserve permissions
-cp --preserve hid-xpadneo/dkms.conf hid-xpadneo/dkms.conf_bck
-cp --preserve hid-xpadneo/src/hid-xpadneo.c hid-xpadneo/src/hid-xpadneo.c_bck
-
-echo "* replacing version string if necessary"
-sed -i 's/PACKAGE_VERSION="@DO_NOT_CHANGE@"/PACKAGE_VERSION="'"$VERSION"'"/g' hid-xpadneo/dkms.conf
-sed -i 's/#define DRV_VER "@DO_NOT_CHANGE@"/#define DRV_VER "'"$VERSION"'"/g' hid-xpadneo/src/hid-xpadneo.c
-
-
 INSTALLED=$(dkms status 2>/dev/null | grep '^hid-xpadneo,' 2>/dev/null | sed -E 's/^hid-xpadneo, ([0-9]+.[0-9]+.[0-9]+).*installed/\1/')
 
 if [[ -z "$INSTALLED" ]]; then
 
     echo "* copying module into /usr/src"
     cp --recursive "$PWD/hid-xpadneo/" "/usr/src/hid-xpadneo-$VERSION"
+
+    echo "* setting version strings (if necessary)"
+    sed -i 's/PACKAGE_VERSION="@DO_NOT_CHANGE@"/PACKAGE_VERSION="'"$VERSION"'"/g' "/usr/src/hid-xpadneo-$VERSION/dkms.conf"
+    sed -i 's/#define DRV_VER "@DO_NOT_CHANGE@"/#define DRV_VER "'"$VERSION"'"/g' "/usr/src/hid-xpadneo-$VERSION/src/hid-xpadneo.h"
 
     echo "* adding module to DKMS"
     dkms add -m hid-xpadneo -v "$VERSION"
@@ -37,7 +31,3 @@ else
     echo "already installed!"
 
 fi
-
-# restore original files
-mv hid-xpadneo/dkms.conf_bck hid-xpadneo/dkms.conf
-mv hid-xpadneo/src/hid-xpadneo.c_bck hid-xpadneo/src/hid-xpadneo.c
